+++ frontend/nuxt.config.ts
@@ -22,6 +22,8 @@ export default defineNuxtConfig({
       docsUrl: 'https://github.com/fayouz/rocket-mailer/tree/develop/docs/content',
       changelogUrl: 'https://github.com/fayouz/rocket-mailer/blob/develop/CHANGELOG.md',
       mailpitUrl: '',
+      // Version of the interface: set by the Docker image (NUXT_PUBLIC_APP_VERSION).
+      appVersion: 'dev',
     },
   },
   icon: {